Included a class function for extracting the a file context into the message.
