+++ shipyard-templates/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "shipyard-templates"
-version = "0.6.2a0"
+version = "0.6.2a1"
 description = "Super classes for blueprint development"
 authors = ["wrp801 <wespoulsen@gmail.com>"]
 license = "Apache License 2.0"
